requirements.txt
streamlit
ultralytics
torch
torchvision
matplotlib
pillow
plotly